requirements.txt
aiohttp==3.6.2
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
cachetools==4.0.0
chardet==3.0.4
decorator==4.0.11
Django==3.0.4
ez-setup==0.9
facebook-sdk==3.1.0
google-api-python-client==1.7.11
google-auth==1.11.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
httplib2==0.17.0
idna==2.9
idna-ssl==1.1.0
imageio==2.1.2
InstagramAPI==1.0.2
moviepy==0.2.3.2
multidict==4.7.5
numpy==1.18.1
oauthlib==3.1.0
pandas==1.0.1
Pillow==7.0.0
praw==6.5.1
prawcore==1.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2019.3
requests==2.11.1
requests-oauthlib==1.3.0
requests-toolbelt==0.7.0
rsa==4.0
six==1.14.0
sqlparse==0.3.1
telepot==12.7
tqdm==4.11.2
typing-extensions==3.7.4.1
update-checker==0.16
uritemplate==3.0.1
urllib3==1.25.8
websocket-client==0.57.0
yarl==1.4.2
